+++ integrations/generic_log_pack/__init__.py
@@ -1,0 +1,2 @@
+"""Generic log folder packer."""
+